fixed according to the pull request review #160
